main-crm-processor/backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
openai
langchain
langchain-community
pymongo
presidio-analyzer
pydantic
cors
fastapi-cors